database h2 configs e massa criado

+++ src/main/java/com/sistema/libraryManagement/dto/BookDTO.java
@@ -23,7 +23,7 @@ public class BookDTO {
     private String actor;
 
     @JsonProperty("isbn")
-    private int isbn;
+    private Long isbn;
 
     @JsonProperty("publishingCompany")
     private String publishingCompany;
